% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accuracy}
\alias{accuracy}
\title{This function automatically calculates the accuracy}
\usage{
accuracy(pre_probs, labels, threshold = 0.5)
}
\arguments{
\item{pre_probs}{prediction probability from the model}

\item{labels}{true labels}

\item{threshold}{the threshold to determine the predicted labels,defualt is 0.5}
}
\value{
\item{accuracy}{accuracy of the model}
}
\description{
This function automatically calculates the accuracy.
}
\examples{

# build a logistic regression model
data("test_data")
model<-my_logistic_regression(y~x1+x2,data=test_data)
pre <- predict(model,test_data)
accuracy(pre$prob,test_data$y)

}
\references{
Nick T G, Campbell K M. Logistic regression[J]. Topics in biostatistics, 2007:273-301.
}
\seealso{
\code{\link{print.my_logistic_regression}},\code{\link{plot_roc}},\code{\link{predict.my_logistic_regression}}
}
\author{
Chenxi Zhu
}
