% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_logistic_regression.R
\name{my_logistic_regression.formula}
\alias{my_logistic_regression.formula}
\title{Fitting a logistic regression model}
\usage{
\method{my_logistic_regression}{formula}(formula, data, ...)
}
\arguments{
\item{formula}{an object of class "\code{\link{formula}}"}

\item{data}{an optional data frame, list or environment containing the variables in the model.}

\item{\dots}{For \code{my_logistic_regression}: arguments to be used to form the default control argument if it is not supplied directly.}
}
\value{
\item{formula}{models formula} 
\item{coefficients}{a named vector of coefficients} 
\item{residuals}{models residuals} 
\item{fitted.values}{the fitted predict probability} 
\item{aic}{AIC value (Akaike's An Information Criterion)} 
\item{bic}{BIC value} 
\item{call}{the matched call}
\item{data}{training data}
}
\description{
\code{\link{my_logistic_regression.formula}} is used to fit generalized
linear models.The response variable is required to be a binary variable for
0 and 1.
}
\details{
A typical
predictor has the form \code{response ~ terms} where response is the
(numeric) response vector and \code{terms} is a series of terms which
specifies a linear predictor for \code{response}.
}
\note{
The response variable must be a binary variable of 0 and 1.
}
\examples{

# build a logistic regression model
data("test_data")
model<-my_logistic_regression(y~x1+x2,data=test_data)
print(model)

}
\references{
Nick T G,Campbell K M. Logistic regression[J]. Topics in biostatistics, 2007:273-301.
}
\seealso{
\code{\link{print.my_logistic_regression}},\code{\link{plot_roc}},\code{\link{predict.my_logistic_regression}}
}
\author{
Chenxi Zhu
}
