% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{test_data}
\format{
A data frame containing the following variables:
\describe{
  \item{x1}{The first explanatory variable}
  \item{x2}{The second explanatory variable}
  \item{y}{The response variable}
}
}
\source{
Randomly generated data.
}
\usage{
test_data
}
\description{
This is a sample dataset for testing.
}
\keyword{datasets}
